requirements.txt
Flask
pymysql
python-dotenv
APScheduler
